- Merged super class PartialStreetEdge into child class TemporaryPartialStreetEdge, witch were the only place it was used except for tests. - Simplified the constructors a bit.
