+++ src/main/java/org/opentripplanner/routing/services/notes/StaticStreetNotesSource.java
@@ -7,9 +7,7 @@ import java.util.Set;
 
 import org.opentripplanner.common.model.T2;
 import org.opentripplanner.routing.alertpatch.Alert;
-import org.opentripplanner.routing.core.State;
-import org.opentripplanner.routing.core.TraverseMode;
-import org.opentripplanner.routing.edgetype.PartialStreetEdge;
+import org.opentripplanner.routing.edgetype.TemporaryPartialStreetEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -58,8 +56,8 @@ public class StaticStreetNotesSource implements StreetNotesSource, Serializable 
     @Override
     public Set<MatcherAndAlert> getNotes(Edge edge) {
         /* If the edge is temporary, we look for notes in it's parent edge. */
-        if (edge instanceof PartialStreetEdge) {
-            edge = ((PartialStreetEdge) edge).getParentEdge();
+        if (edge instanceof TemporaryPartialStreetEdge) {
+            edge = ((TemporaryPartialStreetEdge) edge).getParentEdge();
         }
         Set<MatcherAndAlert> maas = notesForEdge.get(edge);
         if (maas == null || maas.isEmpty()) {
